Do not let platform font-engine handle zero-width spaces
WebCore has special handling of zero-width spacing, including using
a glyph with index 0 to represent it if the font doesn't have one. This
handling might get overridden by QRawFont::advancesForGlyphIndexes which
redoes advances for everything not marked as spaces.

This patch makes WebCore mark the zero-width spaces as spaces for
width iteration thereby handling it on the WebCore side instead of by
the platform font-engine.

Task-number: QTBUG-40912
Change-Id: Ida1711d3c65f1954273c786c902ef3ae77a72558
Reviewed-by: Michael Bruning <michael.bruning@digia.com>
Reviewed-by: Pierre Rossi <pierre.rossi@gmail.com>